fix: fixed comments
